Add Save button and table list combo box
Can now save the DataSet.
The table list ComboBox can be used to switch between the tables.
The type ComboBox is updated when a cell is selected so that the user can
see the type.

Still can't load a DataSet from a file.
Still can't delete a column. I also think that it is not possible to
delete a row.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>